+++ src/Products/PloneMeeting/behaviors/overrides.py
@@ -9,8 +9,7 @@
 
 
 from collective.behavior.talcondition.behavior import TALCondition
-from plone import api
-from Products.PageTemplates.Expressions import SecureModuleImporter
+from Products.PloneMeeting.utils import _base_extra_expr_ctx
 
 
 class PMTALCondition(TALCondition):
@@ -19,11 +18,7 @@ class PMTALCondition(TALCondition):
     def complete_extra_expr_ctx(self, extra_expr_ctx):
         """Return extra_expr_ctx, this is made to be overrided."""
         extra_context = super(PMTALCondition, self).complete_extra_expr_ctx(extra_expr_ctx)
-        tool = api.portal.get_tool('portal_plonemeeting')
         context = extra_context.get('context', self.context)
-        cfg = tool.getMeetingConfig(context)
-        extra_context.update({'tool': tool,
-                              'pm_utils': SecureModuleImporter['Products.PloneMeeting.utils'],
-                              'imio_history_utils': SecureModuleImporter['imio.history.utils'],
-                              'cfg': cfg})
+        base_extra_expr_ctx = _base_extra_expr_ctx(context)
+        extra_context.update(base_extra_expr_ctx)
         return extra_context